ft: move helper code to the reader class
